teleconsult-dr-backend-main/requirements.txt
Django==4.2.1
channels==3.0.3
djangorestframework
rest_condition==1.0.3
django-widget-tweaks==1.4.8
django-cors-headers==3.7.0
requests~=2.25.1
PyJWT<=1.7.1
pytz~=2021.1
xlrd==1.2.0
django_extensions
django-crontab
pyotp
django-environ
psycopg2-binary==2.8.6
Pillow==9.4.0